Agregando API factura, pedido, producto

+++ APIProyectoRapi/src/controller/factura.controller.js
@@ -5,23 +5,29 @@ const Factura = require('../model/factura.model')
 facturaController.getFacturas = async (req, res) => {
     try {
         const facturas = await Factura.find({});
-        res.status(200).json(facturas);
+        if(Object.entries(facturas).length === 0){
+            res.status(200).send("No se encontraron facturas en el sistema");
+        }else{
+            res.status(200).json(facturas);
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 1");
+        res.status(500).send("Ocurrió un error al consultar las facturas");
     }
-
 };
 
 
 facturaController.getFactura = async (req, res) => {
     try {
         const factura = await Factura.findOne({ numeroFactura: req.params.numeroFactura });
-
-        res.status(200).json(factura);
+        if(factura === null){
+            res.status(200).send("No se encontró la factura en el sistema");
+        }else{
+            res.status(200).json(factura);
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 2");
+        res.status(500).send("Ocurrió un error al consultar la factura");
     }
 };
 
@@ -40,46 +46,51 @@ facturaController.createFactura = async (req, res) => {
         let _factura = new Factura({numeroFactura, cliente, fecha, numPedido, lugarEntrega, modoPago});
         await _factura.save();
 
-        res.status(201).send("Factura creado satisfactoriamente.");
+        res.status(201).send("Factura creada satisfactoriamente");
 
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 3");
+        res.status(500).send("Ocurrió un error al crear la factura");
     }
 
 };
 
 facturaController.editFactura = async (req, res) => {
     try {
-        const {
-            numeroFactura,    
-            cliente,
-            fecha,
-            numPedido,
-            lugarEntrega,
-            modoPago
-        } = req.body;
+        const factura = await Factura.findOne({ numeroFactura: req.params.numeroFactura });
+        if(factura === null){
+            res.status(200).send("No se encontró la factura en el sistema");
+        }else{
+            const {
+                numeroFactura,    
+                cliente,
+                fecha,
+                numPedido,
+                lugarEntrega,
+                modoPago
+            } = req.body;
 
-
-        await Factura.updateOne({ numeroFactura: req.params.numeroFactura }, {numeroFactura, cliente, fecha, numPedido, lugarEntrega, modoPago});
-
-        res.status(200).send("Actualizacion de factura satisfactoria.");
-
+            await Factura.updateOne({ numeroFactura: req.params.numeroFactura }, {numeroFactura, cliente, fecha, numPedido, lugarEntrega, modoPago});
+            res.status(200).send("Actualizacion de factura satisfactoria.");
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 4");
+        res.status(500).send("Ocurrió un error al actualizar la factura");
     }
 };
 
 facturaController.deleteFactura = async (req, res) => {
     try {
-
-        await Factura.deleteOne({numeroFactura: req.params.numeroFactura });
-        res.status(200).send("Factura eliminado con éxito.");
-
+        const factura = await Factura.findOne({ numeroFactura: req.params.numeroFactura });
+        if(factura === null){
+            res.status(200).send("No se encontró la factura en el sistema");
+        }else{
+            await Factura.deleteOne({numeroFactura: req.params.numeroFactura });
+            res.status(200).send("Factura eliminada con éxito.");
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 5");
+        res.status(500).send("Ocurrió un error al eliminar la factura");
     }
 };
