+++ APIProyectoRapi/src/controller/producto.controller.js
@@ -17,82 +17,88 @@ const Producto = require('../model/producto.model');
 productoController.getProductos = async (req, res) => {
     try {
         const productos = await Producto.find({});
-        res.status(200).json(productos);
+        if(Object.entries(productos).length === 0){
+            res.status(200).send("No se encontraron productos en el sistema");
+        }else{
+            res.status(200).json(productos);
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 1");
+        res.status(500).send("Ocurrió un error al consultar los productos");
     }
-
 };
-
 
 productoController.getProducto = async (req, res) => {
     try {
         const producto = await Producto.findOne({ codigo: req.params.codigo });
-
-        res.status(200).json(producto);
+        if(producto === null){
+            res.status(200).send("No se encontró el producto en el sistema");
+        }else{
+            res.status(200).json(producto);
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 2");
+        res.status(500).send("Ocurrió un error al consultar el producto");
     }
 };
 
-
 productoController.createProducto = async (req, res) => {
     try {
-        const {
-            categoria,
-            nombre,
-            codigo,
-            descripcion,
-            precio,
-            unidades
+        const productoTemp ={
+            categoria: req.body.categoria,
+            nombre: req.body.nombre,
+            codigo: req.body.codigo,
+            descripcion: req.body.descripcion,
+            precio: req.body.precio,
+            unidades: req.body.unidades
+        };
 
-        } = req.body;
-
-        let _producto = new Producto({categoria, nombre, codigo, descripcion, precio, unidades });
+        let _producto = new Producto(productoTemp);
         await _producto.save();
-
         res.status(201).send("Producto creado satisfactoriamente.");
 
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 3");
+        res.status(500).send("Ocurrió un error al crear el producto");
     }
-
 };
 
 productoController.editProducto = async (req, res) => {
     try {
-        const {
-            categoria,
-            nombre,
-            codigo,
-            descripcion,
-            precio,
-            unidades
-        } = req.body;
+        const producto = await Producto.findOne({codigo: req.params.codigo});
+        if(producto === null){
+            res.status(200).send("No se encontró el producto en el sistema");
+        }else{
+            const {
+                categoria,
+                nombre,
+                codigo,
+                descripcion,
+                precio,
+                unidades
+            } = req.body;
 
-
-        await Producto.updateOne({ codigo: req.params.codigo }, { categoria, nombre, codigo, descripcion, precio, unidades });
-
-        res.status(200).send("Actualizacion de producto satisfactoria.");
-
+            await Producto.updateOne({ codigo: req.params.codigo }, { categoria, nombre, codigo, descripcion, precio, unidades });
+            res.status(201).send("Actualizacion de producto satisfactoria");
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 4");
+        res.status(500).send("Ocurrió un error al editar el producto");
     }
 };
 
 productoController.deleteProducto = async (req, res) => {
     try {
-
-        await Producto.deleteOne({ codigo: req.params.codigo });
-        res.status(200).send("Producto eliminado con éxito.");
-
+        const producto = await Producto.findOne({codigo: req.params.codigo});
+        if(producto === null){
+            res.status(200).send("No se encontró el producto en el sistema");
+        }else{
+            await Producto.deleteOne({ codigo: req.params.codigo });
+            res.status(200).send("Producto eliminado con éxito.");
+        }
     } catch (error) {
         console.log(error);
-        res.status(400).send("Ocurrió un error en la operacion 5");
+        res.status(500).send("Ocurrió un error al eliminar el producto");
     }
 };
